# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic restfull api for reddit-clone backend.",
        "title": "Reddit-Clone API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/post/{POST_ID}": {
            "get": {
                "description": "Get information on a specific post by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "getting-posts"
                ],
                "summary": "Get a certain post",
                "operationId": "get-post-by-id",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "Post uuid",
                        "name": "POST_ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post successfully received",
                        "schema": {
                            "$ref": "#/definitions/posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad post id",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "404": {
                        "description": "No posts with the provided id were found",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a specific post by its id",
                "tags": [
                    "managing-posts"
                ],
                "summary": "Delete a post",
                "operationId": "delete-post",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "Post uuid",
                        "name": "POST_ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "400": {
                        "description": "Bad post id",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "404": {
                        "description": "No posts with the provided id were found",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
        "/post/{POST_ID}/downvote": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Decrease post rating by 1 vote",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "voting-posts"
                ],
                "summary": "Vote down on a post",
                "operationId": "downvote-post",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "Post uuid",
                        "name": "POST_ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully downvoted",
                        "schema": {
                            "$ref": "#/definitions/posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad post id",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "404": {
                        "description": "No posts with the provided id were found",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
        "/post/{POST_ID}/unvote": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Withdraw your vote from the post",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "voting-posts"
                ],
                "summary": "Cancel your vote",
                "operationId": "unvote-post",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "Post uuid",
                        "name": "POST_ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully unvoted",
                        "schema": {
                            "$ref": "#/definitions/posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad post id",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "404": {
                        "description": "No posts with the provided id were found",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
        "/post/{POST_ID}/upvote": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Increase post rating by 1 vote",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "voting-posts"
                ],
                "summary": "Vote up on a post",
                "operationId": "upvote-post",
                "parameters": [
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "Post uuid",
                        "name": "POST_ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully upvoted",
                        "schema": {
                            "$ref": "#/definitions/posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad post id",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "404": {
                        "description": "No posts with the provided id were found",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
        "/posts/": {
            "get": {
                "description": "Get a list of posts of all users and threads",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "getting-posts"
                ],
                "summary": "Get all posts",
                "operationId": "get-all-posts",
                "responses": {
                    "200": {
                        "description": "Posts successfully received",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/posts.Post"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a post of a specific type, category, and content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "managing-posts"
                ],
                "summary": "Create a post",
                "operationId": "create-post",
                "parameters": [
                    {
                        "description": "Post data",
                        "name": "post_payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/posts.PostPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Post successfully created",
                        "schema": {
                            "$ref": "#/definitions/posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad payload"
                    },
                    "422": {
                        "description": "Bad content",
                        "schema": {
                            "$ref": "#/definitions/errs.ComplexErrArr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
        "/posts/{CATEGORY_NAME}": {
            "get": {
                "description": "Get all posts belonging to a certain category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "getting-posts"
                ],
                "summary": "Get posts by category",
                "operationId": "get-posts-by-category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category name",
                        "name": "CATEGORY_NAME",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts successfully received",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/posts.Post"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad category(doesn't exist)",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
        "/posts/{POST_ID}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Leave a comment under a certain post",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "commenting-posts"
                ],
                "summary": "Comment on the post",
                "operationId": "add-comment",
                "parameters": [
                    {
                        "description": "Comment data",
                        "name": "comment_payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/posts.Comment"
                        }
                    },
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "Post uuid",
                        "name": "POST_ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comment successfully left",
                        "schema": {
                            "$ref": "#/definitions/posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad payload",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "404": {
                        "description": "No posts with the provided id were found",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    },
                    "422": {
                        "description": "Bad content",
                        "schema": {
                            "$ref": "#/definitions/errs.ComplexErrArr"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/errs.SimpleErr"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errs.ComplexErrArr": {
            "description": "ComplexErrArr is an array of ComplexErr returned in case of a non-obvious error",
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/errs.ComplexErr"
                    }
                }
            }
        },
        "errs.SimpleErr": {
            "description": "SimpleErr stores a brief description of an error",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Any type"
                }
            }
        },
...
        "posts.Comment": {
            "description": "Comment contains the text of the comment on Post",
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "minLength": 4,
                    "example": "Some comment body example"
                }
            }
        },
...
        "posts.PostPayload": {
            "description": "PostPayload contains the necessary information to create a post",
            "type": "object",
            "properties": {
                "category": {
                    "description": "Number of the category to which the Post belongs",
                    "allOf": [
                        {
                            "$ref": "#/definitions/posts.PostCategory"
                        }
                    ],
                    "example": 0
                },
                "text": {
                    "description": "Content of the Post",
                    "type": "string",
                    "minLength": 4,
                    "example": "Awesome text"
                },
                "title": {
                    "type": "string",
                    "example": "Awesome title"
                },
                "type": {
                    "description": "link or text",
                    "allOf": [
                        {
                            "$ref": "#/definitions/posts.PostType"
                        }
                    ]
                },
                "url": {
                    "type": "string",
                    "example": "http://localhost:8080/"
                }
            }
        },
        "posts.PostType": {
            "description": "PostType is an integer(0 or 1) representing the type of the Post",
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "WithLink",
                "WithText"
            ]
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}